fix(cli): keep node:fs out of the XLE browser bundle
XLEPanel → @xds/cli/xle → validate.mjs imported levenshteinDistance from
string-utils.mjs, which dynamically imports node:fs/path/url for component
search. Native ESM only evaluates those on call (so the standalone
playground worked), but webpack statically bundles dynamic imports too and
fails: 'UnhandledSchemeError: Reading from node:fs is not handled'.

Extracts the pure levenshteinDistance into lib/levenshtein.mjs; validate.mjs
imports it from there, and string-utils.mjs re-exports it (consumers
unchanged). The whole lib/xle browser chain is now free of node: schemes, so
the docsite playground tab bundles cleanly. Standalone serve.mjs updated to
serve+rewrite the new path. 57 xle tests green.

(cherry picked from commit 928fc82c7784e33b9834f41d006e91fb43243e11)
