+++ packages/cli/playground/serve.mjs
@@ -63,14 +63,14 @@ const server = createServer(async (req, res) => {
     }
 
     if (path.startsWith('/xle/')) {
-      // Serve the pure language modules (and string-utils) as ESM.
+      // Serve the pure language modules (+ levenshtein from lib/) as ESM.
       const file = path.slice('/xle/'.length);
-      const abs = file === 'string-utils.mjs'
-        ? join(CLI_SRC, 'lib/string-utils.mjs')
+      const abs = file === 'levenshtein.mjs'
+        ? join(CLI_SRC, 'lib/levenshtein.mjs')
         : join(CLI_SRC, 'lib/xle', file);
       if (!abs.startsWith(CLI_SRC) || !file.endsWith('.mjs')) return send(res, 403, 'text/plain', 'forbidden');
-      // validate.mjs imports '../string-utils.mjs' — rewrite to a flat /xle path.
-      const body = readFileSync(abs, 'utf-8').replace(/from '\.\.\/string-utils\.mjs'/g, "from './string-utils.mjs'");
+      // validate.mjs imports '../levenshtein.mjs' — rewrite to a flat /xle path.
+      const body = readFileSync(abs, 'utf-8').replace(/from '\.\.\/levenshtein\.mjs'/g, "from './levenshtein.mjs'");
       return send(res, 200, 'text/javascript; charset=utf-8', body);
     }
