+++ packages/cli/src/lib/string-utils.mjs
@@ -4,20 +4,11 @@
  * @file String utilities — fuzzy matching and semantic search for component names
  */
 
-export function levenshteinDistance(a, b) {
-  const m = a.length, n = b.length;
-  const dp = Array.from({length: m + 1}, () => Array(n + 1).fill(0));
-  for (let i = 0; i <= m; i++) dp[i][0] = i;
-  for (let j = 0; j <= n; j++) dp[0][j] = j;
-  for (let i = 1; i <= m; i++) {
-    for (let j = 1; j <= n; j++) {
-      dp[i][j] = a[i-1] === b[j-1]
-        ? dp[i-1][j-1]
-        : 1 + Math.min(dp[i-1][j], dp[i][j-1], dp[i-1][j-1]);
-    }
-  }
-  return dp[m][n];
-}
+// levenshteinDistance lives in its own pure module so browser-bundled code
+// (the XLE/XLO language) can use it without this file's dynamic node:fs/path
+// imports landing in the webpack graph. Re-exported for existing consumers.
+import {levenshteinDistance} from './levenshtein.mjs';
+export {levenshteinDistance} from './levenshtein.mjs';
 
 /**
  * Find the closest component names to a given (possibly misspelled) name.
